+++ src/control/ros/RosJointCommandExecutor.cpp
@@ -1,0 +1,15 @@
+#include "aikido/control/ros/RosJointCommandExecutor.hpp"
+
+namespace aikido {
+namespace control {
+namespace ros {
+
+template class RosJointCommandExecutor<ExecutorType::POSITION>;
+
+template class RosJointCommandExecutor<ExecutorType::VELOCITY>;
+
+template class RosJointCommandExecutor<ExecutorType::EFFORT>;
+
+} // namespace ros
+} // namespace control
+} // namespace aikido