+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Control
 
   * Fixed bug in RosTrajectoryExecutor [#596](https://github.com/personalrobotics/aikido/pull/596)
+  * Added generic and position/velocity/effort Executor types [#602](https://github.com/personalrobotics/aikido/pull/602)
 
 * Robot
   * Simplify Robot API [#593](https://github.com/personalrobotics/aikido/pull/593)